undid legacy package. will do it in separate bracnch
